+++ src/main/java/mil/nga/geopackage/GeoPackage.java
@@ -2,6 +2,7 @@ package mil.nga.geopackage;
 
 import java.sql.ResultSet;
 
+import mil.nga.geopackage.attributes.AttributesDao;
 import mil.nga.geopackage.core.contents.Contents;
 import mil.nga.geopackage.features.columns.GeometryColumns;
 import mil.nga.geopackage.features.user.FeatureDao;
@@ -19,6 +20,7 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Feature DAO from Geometry Columns
 	 *
 	 * @param geometryColumns
+	 *            geometry columns
 	 * @return feature dao
 	 */
 	public FeatureDao getFeatureDao(GeometryColumns geometryColumns);
@@ -27,6 +29,7 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Feature DAO from Contents
 	 *
 	 * @param contents
+	 *            contents
 	 * @return feature dao
 	 */
 	public FeatureDao getFeatureDao(Contents contents);
@@ -35,6 +38,7 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Feature DAO from a table name
 	 *
 	 * @param tableName
+	 *            table name
 	 * @return feature dao
 	 */
 	public FeatureDao getFeatureDao(String tableName);
@@ -43,6 +47,7 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Tile DAO from Tile Matrix Set
 	 *
 	 * @param tileMatrixSet
+	 *            tile matrix set
 	 * @return tile dao
 	 */
 	public TileDao getTileDao(TileMatrixSet tileMatrixSet);
@@ -51,6 +56,7 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Tile DAO from Contents
 	 *
 	 * @param contents
+	 *            contents
 	 * @return tile dao
 	 */
 	public TileDao getTileDao(Contents contents);
@@ -59,9 +65,30 @@ public interface GeoPackage extends GeoPackageCore {
 	 * Get a Tile DAO from a table name
 	 *
 	 * @param tableName
+	 *            table name
 	 * @return tile dao
 	 */
 	public TileDao getTileDao(String tableName);
+
+	/**
+	 * Get an Attributes DAO from Contents
+	 * 
+	 * @param contents
+	 *            contents
+	 * @return attributes dao
+	 * @since 1.2.1
+	 */
+	public AttributesDao getAttributesDao(Contents contents);
+
+	/**
+	 * Get an Attributes DAO from a table name
+	 * 
+	 * @param tableName
+	 *            table name
+	 * @return attributes dao
+	 * @since 1.2.1
+	 */
+	public AttributesDao getAttributesDao(String tableName);
 
 	/**
 	 * Perform a query on the database